refactor(auth): move `AuthenticatedScreen` from LoginWithEmailScreen.kt to RootAuthScreen.kt
