+++ feature/authentication/src/main/java/com/crisiscleanup/feature/authentication/ui/LoginWithEmailScreen.kt
@@ -246,42 +246,3 @@ private fun LoginWithEmailScreen(
         )
     }
 }
-
-@Composable
-internal fun AuthenticatedScreen(
-    authState: AuthenticationState,
-    closeAuthentication: () -> Unit = {},
-    viewModel: AuthenticationViewModel = hiltViewModel(),
-) {
-    val translator = LocalAppTranslator.current
-
-    Text(
-        modifier = fillWidthPadded.testTag("authedProfileAccountInfo"),
-        text = translator("info.account_is")
-            .replace("{full_name}", authState.accountData.fullName)
-            .replace("{email_address}", authState.accountData.emailAddress),
-    )
-
-    val authErrorMessage by viewModel.errorMessage
-    ConditionalErrorMessage(authErrorMessage)
-
-    val isNotBusy by viewModel.isNotAuthenticating.collectAsStateWithLifecycle()
-
-    BusyButton(
-        modifier = fillWidthPadded.testTag("authedProfileLogoutBtn"),
-        onClick = viewModel::logout,
-        enabled = isNotBusy,
-        text = translator("actions.logout"),
-        indicateBusy = !isNotBusy,
-    )
-
-    LinkAction(
-        "actions.back",
-        modifier = Modifier
-            .listItemPadding()
-            .testTag("authedProfileDismissBtn"),
-        arrangement = Arrangement.Start,
-        enabled = isNotBusy,
-        action = closeAuthentication,
-    )
-}
